+++ resources/views/partials/integrations/alertas.blade.php
@@ -1,0 +1,3 @@
+@if (\MyLaravelTools\Panel\Support\PanelIntegrations::alertasEnabled())
+    {!! view('alertas::components.contenedor')->render() !!}
+@endif